requirements.txt
Flask~=2.1.2
Pillow~=9.1.1
boto3~=1.24.5
retina-face==0.0.10
numpy~=1.22.4